+++ modules/Food_Service/Students/ServeMenus.php
@@ -124,7 +124,10 @@ if ( UserStudentID() && ! $_REQUEST['modfunc'] )
 
 	echo '</form>';
 
-	DrawHeader( NoInput( $student['FULL_NAME'], '&nbsp;' . $student['STUDENT_ID'] ), '', NoInput( red( $student['BALANCE'] ), _( 'Balance' ) ) );
+	DrawHeader(
+		NoInput( $student['FULL_NAME'], $student['STUDENT_ID'] ),
+		NoInput( red( $student['BALANCE'] ), _( 'Balance' ) )
+	);
 
 	if ( $student['BALANCE'] != '' )
 	{