+++ modules/Food_Service/Students/Statements.php
@@ -70,7 +70,7 @@ if ( UserStudentID() && ! $_REQUEST['modfunc'] )
 	echo '</form>';
 
 	DrawHeader(
-		NoInput( $student['FULL_NAME'], '&nbsp;' . $student['STUDENT_ID'] ),
+		NoInput( $student['FULL_NAME'], $student['STUDENT_ID'] ),
 		NoInput( red( $student['BALANCE'] ), _( 'Balance' ) )
 	);