+++ modules/Food_Service/Users/ServeMenus.php
@@ -79,7 +79,10 @@ if ( UserStaffID()
 
 	echo '</form>';
 
-	DrawHeader( NoInput( $staff['FULL_NAME'], '&nbsp;' . $staff['STAFF_ID'] ), '', NoInput( red( $staff['BALANCE'] ), _( 'Balance' ) ) );
+	DrawHeader(
+		NoInput( $staff['FULL_NAME'], $staff['STAFF_ID'] ),
+		NoInput( red( $staff['BALANCE'] ), _( 'Balance' ) )
+	);
 
 	if ( $staff['ACCOUNT_ID'] && $staff['BALANCE'] != '' )
 	{