+++ modules/Food_Service/Users/Transactions.php
@@ -71,7 +71,7 @@ if ( UserStaffID()
 //FJ fix bug no balance
 	//	DrawHeader(NoInput($staff['FULL_NAME'],'&nbsp;'.$staff['STAFF_ID']),'', NoInput(red($student['BALANCE']),_('Balance')));
 	DrawHeader(
-		NoInput( $staff['FULL_NAME'], '&nbsp;' . $staff['STAFF_ID'] ),
+		NoInput( $staff['FULL_NAME'], $staff['STAFF_ID'] ),
 		NoInput( red( $staff['BALANCE'] ), _( 'Balance' ) )
 	);